Major overhaul of queue and method server
Now uses a "Result" model class to communicate inputs/results
and keep track of timing of a task (e.g., when send and received).

Made wrapper classes for the queues. Makes it easier to create
the input/output pairs for clients and server, and makes sure
that they are in the right direction.