Bug fix: Do not serialize errors on return to user
<deep breath>

I was catching an edge case where we used a thread-pool executor
to launch task requests. When tasks crash on the thread pool executor,
it does not re-serialize the data. The error handling thread
uses the pointer to the same "result object" as what is used in by
the thread pool executor, which means that the result object will
be unserialized. We are only using the thread pool executor for debug
purposes and, so, do not exepct this edge case to happen in practice.
Rather, we should assume that the data in the result object is
still serialized because no other methods should be altering the
result object (as it is being edited in a separate process).