Colmena Value Server Integration
Usage of the value server is now integrated into the client queues class. The client queue now accepts an optional threshold argument that, if provided, will be used to add all objects with a greater size than the threshold to the value server (and the result object will contain proxies instead). A separate redis server hostname and port can also be specified for the value server