Proxy after serialization and async resolve bug fix. Closes #22
- input/output serialization now done on a per-value basis. I.e. instead
of serializing the input tuple (*args, **kwargs), each value in *args
and **kwargs are independently serialized.
- Object proxy thresholding is now done on the size of the serialized
object to provide better support for objects like ML models where
`sys.getsizeof(model)` may not be correct.
- Added `is_serialized` flag to `to_proxy()` and `ValueServer.set()` to
signal that the passed object is already serialized and can be put
straight into the value server.
- Fixed a bug where calling `async_resolve()` on a Proxy would resolve
the object in order to check if it needs to resolve itself.
- Added test case for asynchronously resolving proxies.