Better value server serialization
Updated to work with ProxyStore v0.2.0 and froze ProxyStore version
dependency to v0.2.* (because the ProxyStore API may change in the
future).

ProxyStore v0.2.0 adds some useful features for avoiding the double
serialization problem introduced by PR #28.
In particular, custom `Factory` classes can be created and passed to
ProxyStore's `proxy()` function.
This commit adds a custom `ColmenaSerializationFactory` that uses the
Colmena utilities for deserialization when resolving a proxy.

ProxyStore v0.2.0 also adds support for auto-evicting proxies from Redis
which we do with the result values of tasks (because they are only ever
used once in the value server from transportation from a worker back to
the thinker).