Fix Result.json() resolving proxies
Pydantic's `BaseModel.json()` inspects all of the elements inside the object
(a `Result` in Colmena's case) and inadvertantly resolves proxies in the
process. To fix this, we request a dict using Pydantics provided `.dict()`
method but exclude the 'inputs' and 'value' attributes of the `Result`. Then
we use `json.dumps` to serialize a custom dict and use the
`proxy_json_encoder` to skip serializing proxies that are unresolved.