Upgrade to ProxyStore v0.4.0
Most of the code changes are fairly straightforward updates import paths
and replaced uses of the deprecated `init_store` with
initialize/register store.

I factored out some code from the Result object to `get_store` to make
the code a bit more readable and add unit tests in a future commit.

The one quirky change here is that ProxyStore keys are now NamedTuples
rather than strings. This matters where the Proxy stats were saved to
the Result as a JSON dict with the ProxyStore key as the JSON keys. I've
just casted the NamedTuple to a string to resolve this. This will still
maintain the uniqueness of the stat keys, and may have the additional
benefit of encoding more information (like the Store type the key came
from).