Upgrade to Proxystore v0.4.0 (#85)
* Upgrade to ProxyStore v0.4.0

Most of the code changes are fairly straightforward updates import paths
and replaced uses of the deprecated `init_store` with
initialize/register store.

I factored out some code from the Result object to `get_store` to make
the code a bit more readable and add unit tests in a future commit.

The one quirky change here is that ProxyStore keys are now NamedTuples
rather than strings. This matters where the Proxy stats were saved to
the Result as a JSON dict with the ProxyStore key as the JSON keys. I've
just casted the NamedTuple to a string to resolve this. This will still
maintain the uniqueness of the stat keys, and may have the additional
benefit of encoding more information (like the Store type the key came
from).

* Add unit tests for colmena.proxy

This module now has 100% code coverage from these tests.

* Dynamically resolve str: Store mapping

Allows us to get rid of the static STORES mapping of strings to Stores
used to make `proxystore_type` JSONable. Now we store the fully
qualified path (e.g., proxystore.store.local.LocalStore) and dynamically
import `LocalStore` using that path when needed.