Migrate Colmena to Pydantic V2 (#141)
* Migrate to pydantic v2

* Clean up test suite warnings

* Fix "Bad file descriptor" error in queues test

This is caused because a file descriptor being closed twice. The file
descriptor, in this case, was a socket opened by the Connection ojects
of the Pipe created by the PipeQueues class. This specific test pickles
and unpickles the PipeQueues, causing the current process to hold two
Connection objects referencing the same socket file descriptor. Once the
Connection objects are garbage collected at the end of the test, the
destructor tries to close the file descriptor, succeeding the first time
and failing the second time.

The solution is to unpickle the PipeQueues in a separate process.

* Support ProxyStore v0.7